Remove bot.ts
